requirements.txt
# python==3.7.4
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
docopt==0.6.2
idna==2.9
num2words==0.5.10
numpy==1.18.2
pandas==1.0.3
pdfkit==0.6.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
wkhtmltopdf==0.2
xlrd==1.2.0
XlsxWriter==1.1.8
